# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Elma API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Register",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Registration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/accept": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Accept",
        "operationId": "accept_api_auth_accept_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/organizations/{org_id}/agents": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Listing",
        "operationId": "listing_api_organizations__org_id__agents_get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Agent"
                  },
                  "title": "Response Listing Api Organizations  Org Id  Agents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Create",
        "operationId": "create_api_organizations__org_id__agents_post",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/organizations/{org_id}/agents/{agent_id}": {
      "put": {
        "tags": [
          "Agents"
        ],
        "summary": "Update",
        "operationId": "update_api_organizations__org_id__agents__agent_id__put",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          },
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/organizations/{org_id}/agents/{agent_id}/publication": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Publication",
        "operationId": "publication_api_organizations__org_id__agents__agent_id__publication_get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          },
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Publication"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Publication Api Organizations  Org Id  Agents  Agent Id  Publication Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Agents"
        ],
        "summary": "Publish",
        "operationId": "publish_api_organizations__org_id__agents__agent_id__publication_put",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          },
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Agent Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Publication"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/public/agents/{org_slug}/{agent_slug}": {
      "get": {
        "tags": [
          "Sharing"
        ],
        "summary": "Public Agent By Slug",
        "operationId": "public_agent_by_slug_api_public_agents__org_slug___agent_slug__get",
        "parameters": [
          {
            "name": "org_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Slug"
            }
          },
          {
            "name": "agent_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Agent Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicAgent"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AcceptInvite": {
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 200,
            "minLength": 20,
            "title": "Token"
          },
          "password": {
            "type": "string",
            "maxLength": 200,
            "minLength": 10,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "token",
          "password"
        ],
        "title": "AcceptInvite"
      },
...
      "Agent": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "instruction": {
            "type": "string",
            "title": "Instruction"
          },
          "kb_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Kb Ids"
          },
          "form_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Form Ids",
            "default": []
          },
          "config": {
            "$ref": "#/components/schemas/AgentConfig"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "publication_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publication Id"
          },
          "published": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "slug",
          "instruction",
          "kb_ids",
          "config",
          "version"
        ],
        "title": "Agent"
      },
      "AgentConfig": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "const": 1,
            "title": "Schema Version",
            "default": 1
          },
          "llm": {
            "type": "string",
            "enum": [
              "mistral-small-latest",
              "mistral-large-latest"
            ],
            "title": "Llm",
            "default": "mistral-small-latest"
          },
          "stt_provider": {
            "type": "string",
            "enum": [
              "mistral",
              "deepgram"
            ],
            "title": "Stt Provider",
            "default": "mistral"
          },
          "stt": {
            "type": "string",
            "enum": [
              "voxtral-mini-latest",
              "flux-general-multi"
            ],
            "title": "Stt",
            "default": "voxtral-mini-latest"
          },
          "tts_provider": {
            "type": "string",
            "enum": [
              "mistral",
              "cartesia"
            ],
            "title": "Tts Provider",
            "default": "mistral"
          },
          "tts": {
            "type": "string",
            "enum": [
              "voxtral-mini-tts-latest",
              "sonic-3.6"
            ],
            "title": "Tts",
            "default": "voxtral-mini-tts-latest"
          },
          "voice": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Voice",
            "default": "en_paul_neutral"
          },
          "language": {
            "type": "string",
            "enum": [
              "auto",
              "English",
              "German",
              "Russian"
            ],
            "title": "Language",
            "default": "auto"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AgentConfig"
      },
      "AgentInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 180,
            "title": "Description",
            "default": ""
          },
          "instruction": {
            "type": "string",
            "maxLength": 12000,
            "minLength": 1,
            "title": "Instruction"
          },
          "kb_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Kb Ids"
          },
          "form_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Form Ids"
          },
          "config": {
            "$ref": "#/components/schemas/AgentConfig"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "instruction"
        ],
        "title": "AgentInput"
      },
...
      "Credentials": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 200,
            "minLength": 10,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "Credentials"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublicAgent": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "publication_id": {
            "type": "string",
            "format": "uuid",
            "title": "Publication Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "publication_id"
        ],
        "title": "PublicAgent"
      },
      "Publication": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "origins": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Origins"
          },
          "embed": {
            "type": "string",
            "title": "Embed"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "enabled",
          "origins",
          "embed",
          "url"
        ],
        "title": "Publication"
      },
      "PublicationInput": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "origins": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Origins"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "enabled",
          "origins"
        ],
        "title": "PublicationInput"
      },
      "Registration": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 200,
            "minLength": 10,
            "title": "Password"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "email",
          "password",
          "name"
        ],
        "title": "Registration"
      },
...
      "User": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "email"
        ],
        "title": "User"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}